🧹 Extract health.py into health/ package
Extracted the large monolithic health.py file into a modular health/
package containing specialized files for models, abstract base classes,
check implementations, registry logic, endpoint routing, and diagnostics.
