+++ packages/workhorse/src/settler_workhorse/health/checks.py
@@ -1,92 +1,11 @@
-"""Comprehensive health checks and monitoring for the Settler workhorse.
-
-Provides health check endpoints, dependency status, readiness/liveness
-probes, and system diagnostics for production operations.
-"""
-
 import asyncio
 import inspect
-import os
-import platform
-
-try:
-    import resource
-except ImportError:
-    resource = None  # Unix-only module
-import sys
 import time
-from abc import ABC, abstractmethod
 from collections.abc import Callable
-from dataclasses import dataclass, field
-from datetime import UTC, datetime
-from enum import Enum
 from typing import Any
 
-
-class HealthStatus(Enum):
-    """Health check status values."""
-
-    HEALTHY = "healthy"
-    DEGRADED = "degraded"
-    UNHEALTHY = "unhealthy"
-    UNKNOWN = "unknown"
-
-
-@dataclass
-class HealthCheckResult:
-    """Result of a single health check."""
-
-    name: str
-    status: HealthStatus
-    message: str
-    details: dict[str, Any] = field(default_factory=dict)
-    duration_ms: float = 0.0
-    timestamp: datetime = field(default_factory=lambda: datetime.now(UTC))
-
-    @property
-    def is_healthy(self) -> bool:
-        """Return True when the check status is healthy."""
-        return self.status == HealthStatus.HEALTHY
-
-
-@dataclass
-class SystemHealth:
-    """Overall system health summary."""
-
-    status: HealthStatus
-    checks: list[HealthCheckResult]
-    timestamp: datetime = field(default_factory=lambda: datetime.now(UTC))
-    version: str = "unknown"
-    uptime_seconds: float = 0.0
-
-    @property
-    def healthy_count(self) -> int:
-        """Return the number of healthy checks."""
-        return sum(1 for c in self.checks if c.status == HealthStatus.HEALTHY)
-
-    @property
-    def degraded_count(self) -> int:
-        """Return the number of degraded checks."""
-        return sum(1 for c in self.checks if c.status == HealthStatus.DEGRADED)
-
-    @property
-    def unhealthy_count(self) -> int:
-        """Return the number of unhealthy checks."""
-        return sum(1 for c in self.checks if c.status == HealthStatus.UNHEALTHY)
-
-
-class HealthCheck(ABC):
-    """Abstract base class for health checks."""
-
-    def __init__(self, name: str, critical: bool = True, timeout_seconds: float = 5.0):
-        self.name = name
-        self.critical = critical
-        self.timeout_seconds = timeout_seconds
-
-    @abstractmethod
-    async def check(self) -> HealthCheckResult:
-        """Execute the health check."""
-        pass
+from settler_workhorse.health.base import HealthCheck
+from settler_workhorse.health.models import HealthCheckResult, HealthStatus
 
 
 class DatabaseHealthCheck(HealthCheck):
@@ -477,268 +396,3 @@ class MemoryHealthCheck(HealthCheck):
                 details={"error": str(e)},
                 duration_ms=duration_ms,
             )
-
-
-class HealthCheckRegistry:
-    """Registry for managing and executing health checks.
-
-    Example:
-        registry = HealthCheckRegistry()
-
-        # Register checks
-        registry.register(DatabaseHealthCheck(check_fn=db_ping))
-        registry.register(RedisHealthCheck())
-
-        # Run all checks
-        health = await registry.check_all()
-
-        # Or run specific check
-        db_health = await registry.check("database")
-    """
-
-    def __init__(self, version: str = "unknown"):
-        self.version = version
-        self._checks: dict[str, HealthCheck] = {}
-        self._start_time = time.time()
-
-    def register(self, check: HealthCheck):
-        """Register a health check."""
-        self._checks[check.name] = check
-
-    def unregister(self, name: str):
-        """Unregister a health check."""
-        self._checks.pop(name, None)
-
-    async def check(self, name: str) -> HealthCheckResult | None:
-        """Run a specific health check."""
-        check = self._checks.get(name)
-        if not check:
-            return None
-
-        return await check.check()
-
-    async def check_all(self) -> SystemHealth:
-        """Run all registered health checks."""
-        results = await asyncio.gather(
-            *[check.check() for check in self._checks.values()],
-            return_exceptions=True,
-        )
-
-        checks: list[HealthCheckResult] = []
-        for result in results:
-            if isinstance(result, Exception):
-                checks.append(
-                    HealthCheckResult(
-                        name="unknown",
-                        status=HealthStatus.UNKNOWN,
-                        message=f"Check failed with exception: {str(result)}",
-                    )
-                )
-            else:
-                checks.append(result)
-
-        # Determine overall status
-        critical_unhealthy = any(
-            r.status == HealthStatus.UNHEALTHY
-            and self._checks.get(r.name, HealthCheck("")).critical
-            for r in checks
-        )
-
-        any_unhealthy = any(r.status == HealthStatus.UNHEALTHY for r in checks)
-        any_degraded = any(r.status == HealthStatus.DEGRADED for r in checks)
-
-        if critical_unhealthy:
-            overall_status = HealthStatus.UNHEALTHY
-        elif any_unhealthy or any_degraded:
-            overall_status = HealthStatus.DEGRADED
-        else:
-            overall_status = HealthStatus.HEALTHY
-
-        return SystemHealth(
-            status=overall_status,
-            checks=checks,
-            version=self.version,
-            uptime_seconds=time.time() - self._start_time,
-        )
-
-    def get_check_names(self) -> list[str]:
-        """Get names of all registered checks."""
-        return list(self._checks.keys())
-
-
-class HealthEndpoint:
-    """HTTP endpoint handler for health checks.
-
-    Provides standard Kubernetes-style endpoints:
-    - /health/live - Liveness probe
-    - /health/ready - Readiness probe
-    - /health - Full health check
-
-    Example:
-        registry = HealthCheckRegistry()
-        endpoint = HealthEndpoint(registry)
-
-        # In your web framework
-        @app.get("/health")
-        async def health():
-            return await endpoint.full_check()
-    """
-
-    def __init__(self, registry: HealthCheckRegistry, ready_checks: list[str] | None = None):
-        self.registry = registry
-        self.ready_checks = ready_checks or []
-
-    async def liveness(self) -> tuple[dict[str, Any], int]:
-        """Liveness probe - is the process running?
-
-        Returns simple 200 if process is alive.
-        """
-        return {
-            "status": "alive",
-            "timestamp": datetime.now(UTC).isoformat(),
-        }, 200
-
-    async def readiness(self) -> tuple[dict[str, Any], int]:
-        """Readiness probe - is the service ready to accept traffic?
-
-        Checks critical dependencies only.
-        """
-        health = await self.registry.check_all()
-
-        # Only check critical services for readiness
-        critical_checks = [c for c in health.checks if c.name in self.ready_checks]
-
-        ready = all(c.is_healthy for c in critical_checks)
-
-        status_code = 200 if ready else 503
-
-        return {
-            "ready": ready,
-            "checks": [
-                {
-                    "name": c.name,
-                    "status": c.status.value,
-                    "message": c.message,
-                }
-                for c in critical_checks
-            ],
-            "timestamp": health.timestamp.isoformat(),
-        }, status_code
-
-    async def full_check(self) -> tuple[dict[str, Any], int]:
-        """Full health check with all registered checks.
-
-        Returns detailed health information.
-        """
-        health = await self.registry.check_all()
-
-        status_code = 200 if health.status == HealthStatus.HEALTHY else 503
-        if health.status == HealthStatus.DEGRADED:
-            status_code = 200  # Degraded still serves traffic
-
-        return {
-            "status": health.status.value,
-            "version": health.version,
-            "uptime_seconds": health.uptime_seconds,
-            "summary": {
-                "healthy": health.healthy_count,
-                "degraded": health.degraded_count,
-                "unhealthy": health.unhealthy_count,
-            },
-            "checks": [
-                {
-                    "name": c.name,
-                    "status": c.status.value,
-                    "message": c.message,
-                    "duration_ms": c.duration_ms,
-                    "details": c.details,
-                }
-                for c in health.checks
-            ],
-            "timestamp": health.timestamp.isoformat(),
-        }, status_code
-
-
-class SystemDiagnostics:
-    """System diagnostics and information gathering.
-
-    Provides detailed system information for debugging and monitoring.
-    """
-
-    @staticmethod
-    def get_system_info() -> dict[str, Any]:
-        """Get comprehensive system information."""
-        info = {
-            "platform": {
-                "system": platform.system(),
-                "release": platform.release(),
-                "version": platform.version(),
-                "machine": platform.machine(),
-                "processor": platform.processor(),
-            },
-            "python": {
-                "version": sys.version,
-                "implementation": platform.python_implementation(),
-            },
-            "environment": {
-                "pid": os.getpid(),
-            },
-        }
-
-        # Try to get resource usage
-        try:
-            import psutil
-
-            process = psutil.Process()
-
-            info["resources"] = {
-                "memory_mb": process.memory_info().rss / (1024 * 1024),
-                "cpu_percent": process.cpu_percent(),
-                "threads": process.num_threads(),
-                "open_files": len(process.open_files()),
-                "connections": len(process.connections()),
-            }
-        except Exception:
-            pass
-
-        return info
-
-    @staticmethod
-    def get_resource_limits() -> dict[str, Any]:
-        """Get system resource limits."""
-        if resource is None:
-            return {"error": "Resource limits not available on Windows"}
-
-        try:
-            limits = {
-                "max_open_files": resource.getrlimit(resource.RLIMIT_NOFILE),
-                "max_memory_mb": (
-                    resource.getrlimit(resource.RLIMIT_AS)[0] / (1024 * 1024)
-                    if resource.getrlimit(resource.RLIMIT_AS)[0] > 0
-                    else "unlimited"
-                ),
-                "max_processes": resource.getrlimit(resource.RLIMIT_NPROC),
-            }
-        except Exception:
-            limits = {"error": "Could not retrieve resource limits"}
-
-        return limits
-
-
-# Export public API
-__all__ = [
-    # Enums
-    "HealthStatus",
-    # Classes
-    "HealthCheckResult",
-    "SystemHealth",
-    "HealthCheck",
-    "DatabaseHealthCheck",
-    "RedisHealthCheck",
-    "ExternalAPIHealthCheck",
-    "DiskSpaceHealthCheck",
-    "MemoryHealthCheck",
-    "HealthCheckRegistry",
-    "HealthEndpoint",
-    "SystemDiagnostics",
-]
