+++ packages/workhorse/src/settler_workhorse/health/models.py
@@ -1,0 +1,56 @@
+from dataclasses import dataclass, field
+from datetime import UTC, datetime
+from enum import Enum
+from typing import Any
+
+
+class HealthStatus(Enum):
+    """Health check status values."""
+
+    HEALTHY = "healthy"
+    DEGRADED = "degraded"
+    UNHEALTHY = "unhealthy"
+    UNKNOWN = "unknown"
+
+
+@dataclass
+class HealthCheckResult:
+    """Result of a single health check."""
+
+    name: str
+    status: HealthStatus
+    message: str
+    details: dict[str, Any] = field(default_factory=dict)
+    duration_ms: float = 0.0
+    timestamp: datetime = field(default_factory=lambda: datetime.now(UTC))
+
+    @property
+    def is_healthy(self) -> bool:
+        """Return True when the check status is healthy."""
+        return self.status == HealthStatus.HEALTHY
+
+
+@dataclass
+class SystemHealth:
+    """Overall system health summary."""
+
+    status: HealthStatus
+    checks: list[HealthCheckResult]
+    timestamp: datetime = field(default_factory=lambda: datetime.now(UTC))
+    version: str = "unknown"
+    uptime_seconds: float = 0.0
+
+    @property
+    def healthy_count(self) -> int:
+        """Return the number of healthy checks."""
+        return sum(1 for c in self.checks if c.status == HealthStatus.HEALTHY)
+
+    @property
+    def degraded_count(self) -> int:
+        """Return the number of degraded checks."""
+        return sum(1 for c in self.checks if c.status == HealthStatus.DEGRADED)
+
+    @property
+    def unhealthy_count(self) -> int:
+        """Return the number of unhealthy checks."""
+        return sum(1 for c in self.checks if c.status == HealthStatus.UNHEALTHY)
